% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadApp.R
\name{downloadApp}
\alias{downloadApp}
\alias{downloadServer}
\alias{downloadInput}
\alias{downloadUI}
\title{Download App}
\usage{
downloadApp(addDate = FALSE, showFilename = FALSE)

downloadServer(id, download, addDate = FALSE, showFilename = FALSE)

downloadInput(id)

downloadUI(id)
}
\arguments{
\item{addDate}{add date to filename if `TRUE`}

\item{showFilename}{show filename in UI if `TRUE`}

\item{id}{identifier for shiny reactive}

\item{download}{reactive or reactiveValues object}
}
\description{
Simple app to download plot or table with filename.
The `downloadUI()` function, if used in `ui()`, shows width and height of plot.
See `downloadHideApp()` for app without this function;
the latter also removes the date on filename.
}
