% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPlotApp.R
\name{downloadPlotApp}
\alias{downloadPlotApp}
\alias{downloadPlotServer}
\alias{downloadPlotInput}
\alias{downloadPlotUI}
\title{Download Plot App}
\usage{
downloadPlotApp()

downloadPlotServer(id, download_plot, filename_plot)

downloadPlotInput(id)

downloadPlotUI(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{download_list}{reactiveValues object}
}
\description{
Download Plot App
}
