% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTableApp.R
\name{downloadTableApp}
\alias{downloadTableApp}
\alias{downloadTableServer}
\alias{downloadTableInput}
\alias{downloadTableOutput}
\title{Download App}
\usage{
downloadTableApp()

downloadTableServer(id, download_list)

downloadTableInput(id)

downloadTableOutput(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{download_list}{reactiveValues object}
}
\description{
Download App
}
