% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_null.R
\name{plot_null}
\alias{plot_null}
\title{Null GG Plot}
\usage{
plot_null(msg = "no data")
}
\arguments{
\item{msg}{character message}
}
\description{
Empty plot with optional message.
Useful for shiny apps with plot output.
}
