% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadApp.R, R/downloadHideApp.R
\name{downloadApp}
\alias{downloadApp}
\alias{downloadServer}
\alias{downloadInput}
\alias{downloadUI}
\alias{downloadHideApp}
\title{Download App}
\usage{
downloadApp()

downloadServer(id, download_list, addDate = FALSE)

downloadInput(id)

downloadUI(id)

downloadHideApp()
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{download_list}{reactiveValues object}

\item{addDate}{add date to filename if `TRUE`}
}
\description{
Simple app to download plot or table with filename.
The `downloadUI()` function, if used in `ui()`, shows width and height of plot.
See `downloadHideApp()` for app without this function;
the latter also removes the date on filename.
}
