% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadApp.R
\name{downloadApp}
\alias{downloadApp}
\alias{downloadServer}
\alias{downloadInput}
\alias{downloadUI}
\title{Download App}
\usage{
downloadApp()

downloadServer(id, download_list, addDate = FALSE)

downloadInput(id)

downloadUI(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{download_list}{reactiveValues object}
}
\description{
Download App
}
